[Pet] Guard the pet action downcasts, and land the name query pair
Clears the block on 2bd6a13ee and adds CMSG_PET_NAME_QUERY with its
response.

SECURITY, and the more serious half of this commit. HandlePetAction
downcast to Pet* in fourteen places without checking. The filter above
them constrains players and real pets; a charmed creature or a vehicle
falls through both arms. Real traffic does exactly that: bodies now
fixtured here carry a HIGHGUID_VEHICLE mover with ACT_COMMAND. Pet
declares members past the end of a Creature, so SetSpellOpener() on a
vehicle wrote out of bounds, driven by a client packet. Registering the
opcode in 2bd6a13ee is what made that path reachable, so it belongs here.
The type is now resolved once and every Pet-only call is guarded.

FIXTURE DISCRIMINATION. A review blocked 2bd6a13ee on the ground that its
four bodies carried only two distinct presence masks, so they reproduced
the reader rather than proving it: bits that are always present together
can be permuted undetected. That was correct. The set is now eight bodies
across FIVE masks -- 0x7904, 0x7DAE, 0x7DAF, 0xFEBF, 0xFFBF.

0x7DAE and 0x7DAF are the pair that settles it. They differ in exactly one
mask bit, for the same pet, and the targets differ accordingly, which pins
that bit's position directly.

CORRECTIONS, all mine. The action-3 fixture was labelled "stay"; action 3
is COMMAND_ABANDON, which for a hunter pet unsummons permanently. The
header claimed five bodies across three lengths; it was four across two.
And it claimed every neighbouring movement body reported the position's
middle float exactly; they vary in the fourth decimal and one by more. The
middle slot is still z, but by coordinate band, and the same bands support
y, z, x for the outer pair rather than leaving it undetermined.

COMMAND_MOVE_TO (action 4) is on the wire but was absent from the enum, so
every ground-targeted pet move fell to default and logged at error level.
It is now accepted and ignored.

PET NAME QUERY, both directions. The request packs the pet GUID and the
pet number behind sixteen interleaved presence bits; the response leads
with bit-field lengths, carries the name unterminated, and trails the pet
number as eight bytes. Four request bodies across three masks, and two
responses. The pair is tied together by evidence: one response is the
actual reply to one of the requests, and the pet number decoded from each
side independently is the same.

Targeted suites pass, 130/130.

+++ src/game/Server/WorldSession.h
@@ -1679,7 +1679,7 @@ class WorldSession
         void SendPetitionQueryOpcode(ObjectGuid petitionguid);
 
         // pet
-        void SendPetNameQuery(ObjectGuid guid, uint32 petnumber);
+        void SendPetNameQuery(ObjectGuid guid, uint64 petnumber);
         void SendStablePet(ObjectGuid guid);
         void SendStableResult(uint8 res);
         bool CheckStableMaster(ObjectGuid guid);
